refactor(flights): regroup video production actions
